Move job.py to core.py
