+++ dask_jobqueue/htcondor.py
@@ -5,7 +5,7 @@ import shlex
 import dask
 from distributed.utils import parse_bytes
 
-from .job import JobQueueCluster, Job, job_parameters, cluster_parameters
+from .core import JobQueueCluster, Job, job_parameters, cluster_parameters
 
 logger = logging.getLogger(__name__)
