+++ dask_jobqueue/pbs.py
@@ -4,7 +4,7 @@ import os
 
 import dask
 
-from .job import Job, JobQueueCluster, job_parameters, cluster_parameters
+from .core import Job, JobQueueCluster, job_parameters, cluster_parameters
 
 logger = logging.getLogger(__name__)
