+++ dask_jobqueue/slurm.py
@@ -3,7 +3,7 @@ import math
 
 import dask
 
-from .job import Job, JobQueueCluster, job_parameters, cluster_parameters
+from .core import Job, JobQueueCluster, job_parameters, cluster_parameters
 
 logger = logging.getLogger(__name__)
